Reorganize prism forward modelling into a submodule
Group together prism-related pieces of code inside a new
`harmonica._forward/prisms` module, organized in a similar way as the
one for `ellipsoids`.
